Convert RedisServer into a class and expose it
Now the RedisServer is a proper class and it is exposed in the
environment. This makes it possible for applications using the core gem
to easily access the Redis instance to set/get application specific
stuff (like raw_messages or api_keys). Without this, I was forced to
either monkeypatch the RedisServer module or add raw_messages and
api_keys functionality to the core gem's RedisServer which is certainly
not the best solution.

This also makes stuff much easier to test.