Webapp: remove dependency on the login extension
Currently, the webapp class automatically includes the login
extension, making it available for all subclassed apps.

However, this causes some problems as the Warden config block
inside LT::WebApp::Login sets the base webapp class as the
current rack app for Warden.

The subclassed app is different from the base webapp. Because
of the way Warden invokes actions in the webapp it has been
configured with, this means we're unable to catch login failures
in webapp subclasses.

It is preferrable to directly require the Login extension in
webapp subclasses.